fix(addTraceurRuntime): save target file in the options object instead of global variable
The tests for bundle.js weren't passing because the runtime was being appended to files unrelated to the test. This was because the runtime target was set the first time the es6ify function was being called at a "require" level, es6ify would then need to be re-required to reset the runtime target.