Add some info about add(es6ify.runtime) effects
Adding `es6ify.runtime` to browserify unfortunately breaks the default behavior decribed on its README (https://github.com/substack/node-browserify#btransformopts-tr), that is, applying transform only on the local package files. It took me a bit of time to understand why one of my submodules was getting broken once in the browser; it was simply because traceur adds "use strict" and it was not compatible.

So I suggest we make that fact explicit here. We may also want to talk about it in the first example (?)