IW: Protect against null source maps
There are situations where traceur will return null for the source map
property. For example, when using browserify's exclude flag, a temporary
but empty file is generated. es6ify will blow up on this file because it
always blindly assumes that there will be a sourcmap property. This
commit checks to see if the source map property exists before parsing
it. If there is no source map property, it just returns the source from
the compiled output.